places SELECT buttons to the left side
this tailwind edit, places the select button on the left side so the user can read and select the missing models easily. Before, the UI placed the select button on the all the way to the right, which makes it hard to recognize which select button belongs to the avalible module. Especially on wide screens.